Translog deletion only for remote txlog enabled indices

+++ server/src/main/java/org/opensearch/index/shard/IndexShard.java
@@ -1904,7 +1904,7 @@ public class IndexShard extends AbstractIndexShardComponent implements IndicesCl
                     // Also closing refreshListeners to prevent us from accumulating any more listeners
                     IOUtils.close(engine, globalCheckpointListeners, refreshListeners, pendingReplicationActions);
 
-                    if (deleted && engine != null && isPrimaryMode()) {
+                    if (deleted && engine != null && isPrimaryMode() && isRemoteTranslogEnabled()) {
                         // Translog Clean up
                         assert engine instanceof InternalEngine;
                         ((InternalEngine) engine).translogManager().onDelete();
